it finally works

+++ lab_4/src/pwm.h
@@ -3,6 +3,6 @@
 void initPWMTimer3();
 void initPWMTimer4();
 
-void changeDutyCycle(float dutycycle1,float dutycycle2);
+void changeDutyCycle(double dutycycle1,double dutycycle2);
 void analyzeADC();
 #endif
